Highlights
----------

This merges the MCUboot 1.1.0 release into the OSF development tree.

This commit will be the final fast-forward update in the MCUboot v1.1.0
development branch. The next update have it history rewritten with the
OSF commits on top of the new v1.1.0 tag as upstream baseline commit.

MCUboot is now built in C99 mode.

MCUboot now supports ECDSA signatures in both the Zephyr and Mynewt
ports, as well as the simulator.

The imgtool helper script now supports generating and managing
password-protected private keys. This adds a new dependency on the
Python cryptography module, replacing the Crypto and pyasn1
dependencies. As a result of this change, private keys are now stored
in the PKCS#8 format, rather than the previous raw
format. Additionally, imgtool no longer supports the PKCS1.5
algorithm; this should not be a problem for users, since MCUboot
itself has not supported this algorithm for some time.

Various tweaks, improvements, and fixes were merged to the core image
update code, including a feature macro, MCUBOOT_VALIDATE_SLOT0, for
controlling whether slot 0 is validated each boot.

The Mynewt port uses the bundled tinycrypt and mbedtls libraries.

The Zephyr port now suports a serial bootloader mode, by which
firmware updates can be sent to the bootloader via a console-friendly
UART protocol.

The Zephyr port now has a top-level Kconfig configuration file.

Upstream Changes
----------------

Cryptography:

- 891f4fb Fix EC API usage due to change in mbed-TLS
- b6e0ae6 fix: Add back ECDSA signature support
- 20462a7 fix: RSA: Use 32-byte salt with PSS
- ba05f2a Add asn1 parser of mbed-tls
- c786540 Add ecdsa to build

Documentation:

- a722f5a Describe the tradeoffs of scratch size

Helper scripts:

- f88d9f9 imgtool: Add password support to ECDSA
- 31d29c8 imgtool: Update docs for new key format
- 2c9153a imgtool: ECDSA pad signature
- 1d5bea1 imgtool: Add support for password protected RSA keys
- 47b77c5 imgtool: Fix signing payload type
- 5e7c6dd imgtool: Update RSA code
- d402257 imgtool: Remove PKCS1.5 support
- 2445473 imgtool: Move keys to submodule
- b730e24 assemble.py: Add explicit license declaration
- 1314bf3 imgtool: Add explicit license declarations
- efe67ae Disable script debugging
- 5f9fbcc Only check signed-off-by in PRs
- 6da40d0 Update to use parent commits
- 0b56138 Besides author also check committer
- 5101b0f Try using tail again
- 3118017 Fix review issues

Image update:

- f70e302 Fix warn in overwrite-only with slot0 validation
- 9c25fa7 Avoid using uninitialized memory in split go
- a91c626 Mark non-used functions in overwrite only mode
- a0e1cce Enable ignoring write failures to status area
- fa137fc Add option to disable verifying contents before writes
- ebdc969 Add MCUBOOT_VALIDATE_SLOT0 as feature
- 1a927dd Update bootutil_sig_verify to receive slen as size_t
- aaf767c Fix warning in overwrite only

Miscellaneous:

- 476fd99 Bump to version 1.1.0
- 49045cc Update with 1.1.0 release notes
- 5398604 Remove duplicated copy/paste residue
- eafa1f0 Add NOTICE file for mcuboot from ASF
- eedcc45 Add test with write status area fail and interruption
- b841f0a Add simple status write test without interruption
- 0bccf9d Build in std=c99 mode
- cd5774b Fix for gcc warnings

MyNewt port:

- 07d65c9 Update Mynewt version numbers
- ac18eec Add Mynewt support for bundled tinycrypt + mbedtls

Simulator and testing:

- 59ae522 sim: Remove extraneous `mut`
- 3df2a0e sim: Update some dependencies
- 57c40f7 Fix assert abstraction in sim
- eb95a58 Enable slot0 validation in travis
- f5c895e Add configurable simulation of bad memory writes
- 9b0ee90 Enable capturing asserts inside simulator
- 43685c2 Enable extra crates in simflash
- d32fd64 Fix mbed-tls' walk_dir in build
- 70f1481 Remove duplicated declaration
- 181e54c Update cargo lock
- a9a510b Add ecdsa testing to travis
- 80fde2f Add ecdsa signing support to the simulator
- 92be3fb Add ecdsa signing support
- 8b619bd Update ring+gcc crates and declare sig-ecdsa feature
- f0e5419 Use travis env variable to control check-signed-off-by
- 5b98910 Add travis' validation of "Signed-off-by:" line

Zephyr port:

- 7cc3261 Correct the path of zephyr sample dts file
- 28ee5b0 Add Zephyr support for bundled mbed-tls ASN1 parser
- 8e96b83 enable serial recovery functionality on the zephyr mcuboot
- 614c4e5 Add cboorattr library for zephyr
- 217c4e0 Add tinycbor library for zephyr
- f6f652e Add serial adapter module for support serial communication for zephyr
- 64ad092 Add Kconfig for the zephyr's build of mcuboot

+++ boot/bootutil/src/image_rsa.c
@@ -254,7 +254,7 @@ bootutil_cmp_rsasig(mbedtls_rsa_context *ctx, uint8_t *hash, uint32_t hlen,
 }
 
 int
-bootutil_verify_sig(uint8_t *hash, uint32_t hlen, uint8_t *sig, int slen,
+bootutil_verify_sig(uint8_t *hash, uint32_t hlen, uint8_t *sig, size_t slen,
   uint8_t key_id)
 {
     mbedtls_rsa_context ctx;
